Idea for association syntax within dynamic attributes
Breaks in older (3.0, not sure about 3.1) versions of Rails due to how
association assignment is handled.

Problems: order of block variables is reversed between dynamic
attributes and callbacks. If adding this, I'd prefer to maintain order.
Current implementation handles this by, when one block variable,
treating as evaluator like normal. When two are provided, the instance
becomes the first and the second is the evaluator. Not sure how
confusing that is, though.